rename RightPane to CoursePaneRoot
