+++ src/components/RightPane/CoursePane/CoursePaneRoot.js
