+++ src/components/RightPane/DesktopTabs.js
@@ -2,7 +2,7 @@ import React, { PureComponent } from 'react';
 import { Tab, Tabs, Typography, Paper } from '@material-ui/core';
 import { FormatListBulleted, MyLocation, Search } from '@material-ui/icons';
 import AddedCoursePane from './AddedCourses/AddedCoursePane';
-import RightPane from './CoursePane/RightPane';
+import RightPane from './CoursePane/CoursePaneRoot';
 import RightPaneStore from '../../stores/RightPaneStore';
 import { handleTabChange } from '../../actions/RightPaneActions';
 import UCIMap from './Map/UCIMap';
